Upgrade glob to 10.2.2 (support Node >= 16)
Upgrade the `glob` dependency to its latest version which does not rely anymore on the transitive dependency `inflight`. This is interesting because `inflight` contains a [memory-leak issue](https://github.com/isaacs/inflight/issues/5) which is raised by Security scanners like Checkmarx.

ALso, glob@9 introduced Typescript support, so there is no need for the `@types/glob` package anymore.

As a consequence, the minimum Node version required is 16 because support for older versions was dropped by glob@9.